Reformatting, move _auth out of pages because it doesnt belong there
